Add intro.scala for Scala basics and syntax practice
Created a new Scala file to begin learning and experimenting with basic Scala syntax and features.

Topics covered in this file include:
- Variable and constant declarations
- Functions and methods
- Control structures (if, for, while)
- Basic data types and expressions

This file lays the groundwork for exploring functional programming in Scala, object-oriented concepts, collections, and pattern matching. It will serve as a starting point for building both scripting utilities and scalable applications in Scala.