Add intro.rkt to start learning Racket programming
Created a new Racket file to begin exploring the basics of the Racket language and the functional programming paradigm.

Topics included in this file:
- Defining variables and constants
- Basic arithmetic and expressions
- Function definitions using `define`
- Conditional logic using `if` and `cond`

This file lays the groundwork for more advanced Racket topics such as recursion, higher-order functions, list processing, and macros. It serves as a foundational reference for future exercises and functional programming projects.